refactor: move functions into common
