+++ src/components/molecules/Visualizer/Engine/Cesium/useEngineRef.ts
@@ -1,15 +1,5 @@
 import * as Cesium from "cesium";
-import {
-  Math as CesiumMath,
-  Cartesian3,
-  Ellipsoid,
-  Quaternion,
-  Scene,
-  Matrix3,
-  Cartographic,
-  EllipsoidTerrainProvider,
-  sampleTerrainMostDetailed,
-} from "cesium";
+import { Math as CesiumMath } from "cesium";
 import { useImperativeHandle, Ref, RefObject, useMemo, useRef } from "react";
 import type { CesiumComponentRef } from "resium";
 
@@ -18,7 +8,23 @@ import type { MouseEvents, MouseEvent } from "../ref";
 
 import builtinPrimitives from "./builtin";
 import Cluster from "./Cluster";
-import { getLocationFromScreenXY, flyTo, lookAt, getCamera, getClock } from "./common";
+import {
+  getLocationFromScreenXY,
+  flyTo,
+  lookAt,
+  getCamera,
+  getClock,
+  lookHorizontal,
+  lookVertical,
+  moveForward,
+  moveBackward,
+  moveUp,
+  moveDown,
+  moveLeft,
+  moveRight,
+  moveOverTerrain,
+  flyToGround,
+} from "./common";
 
 export default function useEngineRef(
   ref: Ref<EngineRef>,
@@ -144,113 +150,53 @@ export default function useEngineRef(
       },
       lookHorizontal: amount => {
         const viewer = cesium.current?.cesiumElement;
-        if (!viewer || viewer.isDestroyed()) return;
-        const camera = viewer.scene.camera;
-        const ellipsoid = viewer.scene.globe.ellipsoid;
-        const surfaceNormal = ellipsoid.geodeticSurfaceNormal(camera.position, new Cartesian3());
-        camera.look(surfaceNormal, amount);
+        if (!viewer || viewer.isDestroyed() || !amount) return;
+        lookHorizontal(viewer, amount);
       },
       lookVertical: amount => {
         const viewer = cesium.current?.cesiumElement;
-        if (!viewer || viewer.isDestroyed()) return;
-        const camera = viewer.scene.camera;
-        const ellipsoid = viewer.scene.globe.ellipsoid;
-        const lookAxis = projectVectorToSurface(camera.right, camera.position, ellipsoid);
-        const surfaceNormal = ellipsoid.geodeticSurfaceNormal(camera.position, new Cartesian3());
-        const currentAngle = CesiumMath.toDegrees(
-          Cartesian3.angleBetween(surfaceNormal, camera.up),
-        );
-        const upAfterLook = rotateVectorAboutAxis(camera.up, lookAxis, amount);
-        const angleAfterLook = CesiumMath.toDegrees(
-          Cartesian3.angleBetween(surfaceNormal, upAfterLook),
-        );
-        const friction = angleAfterLook < currentAngle ? 1 : (90 - currentAngle) / 90;
-        camera.look(lookAxis, amount * friction);
+        if (!viewer || viewer.isDestroyed() || !amount) return;
+        lookVertical(viewer, amount);
       },
       moveForward: amount => {
         const viewer = cesium.current?.cesiumElement;
-        if (!viewer || viewer.isDestroyed()) return;
-        const direction = projectVectorToSurface(
-          viewer.scene.camera.direction,
-          viewer.scene.camera.position,
-          viewer.scene.globe.ellipsoid,
-        );
-        viewer.scene.camera.move(direction, amount);
+        if (!viewer || viewer.isDestroyed() || !amount) return;
+        moveForward(viewer, amount);
       },
       moveBackward: amount => {
         const viewer = cesium.current?.cesiumElement;
-        if (!viewer || viewer.isDestroyed()) return;
-        const direction = projectVectorToSurface(
-          viewer.scene.camera.direction,
-          viewer.scene.camera.position,
-          viewer.scene.globe.ellipsoid,
-        );
-        viewer.scene.camera.move(direction, -amount);
+        if (!viewer || viewer.isDestroyed() || !amount) return;
+        moveBackward(viewer, amount);
       },
       moveUp: amount => {
         const viewer = cesium.current?.cesiumElement;
-        if (!viewer || viewer.isDestroyed()) return;
-        const surfaceNormal = viewer.scene.globe.ellipsoid.geodeticSurfaceNormal(
-          viewer.scene.camera.position,
-          new Cartesian3(),
-        );
-        viewer.scene.camera.move(surfaceNormal, amount);
+        if (!viewer || viewer.isDestroyed() || !amount) return;
+        moveUp(viewer, amount);
       },
       moveDown: amount => {
         const viewer = cesium.current?.cesiumElement;
-        if (!viewer || viewer.isDestroyed()) return;
-        const surfaceNormal = viewer.scene.globe.ellipsoid.geodeticSurfaceNormal(
-          viewer.scene.camera.position,
-          new Cartesian3(),
-        );
-        viewer.scene.camera.move(surfaceNormal, -amount);
+        if (!viewer || viewer.isDestroyed() || !amount) return;
+        moveDown(viewer, amount);
       },
       moveLeft: amount => {
         const viewer = cesium.current?.cesiumElement;
-        if (!viewer || viewer.isDestroyed()) return;
-        const direction = projectVectorToSurface(
-          viewer.scene.camera.right,
-          viewer.scene.camera.position,
-          viewer.scene.globe.ellipsoid,
-        );
-        viewer.scene.camera.move(direction, -amount);
+        if (!viewer || viewer.isDestroyed() || !amount) return;
+        moveLeft(viewer, amount);
       },
       moveRight: amount => {
         const viewer = cesium.current?.cesiumElement;
-        if (!viewer || viewer.isDestroyed()) return;
-        const direction = projectVectorToSurface(
-          viewer.scene.camera.right,
-          viewer.scene.camera.position,
-          viewer.scene.globe.ellipsoid,
-        );
-        viewer.scene.camera.move(direction, amount);
+        if (!viewer || viewer.isDestroyed() || !amount) return;
+        moveRight(viewer, amount);
       },
-      moveOverTerrain: async (offset = 10) => {
+      moveOverTerrain: async offset => {
         const viewer = cesium.current?.cesiumElement;
         if (!viewer || viewer.isDestroyed()) return;
-        const camera = viewer.scene.camera;
-        const height = await sampleTerrainHeight(viewer.scene, camera.position);
-        if (height && height !== 0) {
-          const innerCamera = getCamera(viewer);
-          if (innerCamera && innerCamera?.height < height + offset) {
-            camera.moveUp(height + offset - innerCamera.height);
-          }
-        }
+        moveOverTerrain(viewer, offset);
       },
-      flyToGround: async (camera, options, offset = 10) => {
+      flyToGround: async (camera, options, offset) => {
         const viewer = cesium.current?.cesiumElement;
         if (!viewer || viewer.isDestroyed()) return;
-        const cesiumCamera = viewer.scene.camera;
-        const height = await sampleTerrainHeight(viewer.scene, cesiumCamera.position);
-        if (height && height !== 0) {
-          const groundCamera = { ...camera, height: height + offset };
-          cancelCameraFlight.current?.();
-          cancelCameraFlight.current = flyTo(
-            viewer.scene?.camera,
-            { ...getCamera(viewer), ...groundCamera },
-            options,
-          );
-        }
+        flyToGround(viewer, cancelCameraFlight, camera, options, offset);
       },
       onClick: (cb: ((props: MouseEvent) => void) | undefined) => {
         mouseEventCallbacks.current.click = cb;
@@ -303,35 +249,4 @@ export default function useEngineRef(
   useImperativeHandle(ref, () => e, [e]);
 
   return e;
-}
-
-function projectVectorToSurface(vector: Cartesian3, position: Cartesian3, ellipsoid: Ellipsoid) {
-  const surfaceNormal = ellipsoid.geodeticSurfaceNormal(position, new Cartesian3());
-  const magnitudeOfProjectionOnSurfaceNormal = Cartesian3.dot(vector, surfaceNormal);
-  const projectionOnSurfaceNormal = Cartesian3.multiplyByScalar(
-    surfaceNormal,
-    magnitudeOfProjectionOnSurfaceNormal,
-    new Cartesian3(),
-  );
-  return Cartesian3.subtract(vector, projectionOnSurfaceNormal, new Cartesian3());
-}
-
-function rotateVectorAboutAxis(vector: Cartesian3, rotateAxis: Cartesian3, rotateAmount: number) {
-  const quaternion = Quaternion.fromAxisAngle(rotateAxis, -rotateAmount, new Quaternion());
-  const rotation = Matrix3.fromQuaternion(quaternion, new Matrix3());
-  const rotatedVector = Matrix3.multiplyByVector(rotation, vector, vector.clone());
-  return rotatedVector;
-}
-
-async function sampleTerrainHeight(
-  scene: Scene,
-  position: Cartesian3,
-): Promise<number | undefined> {
-  const terrainProvider = scene.terrainProvider;
-  if (terrainProvider instanceof EllipsoidTerrainProvider) return 0;
-
-  const [sample] = await sampleTerrainMostDetailed(terrainProvider, [
-    Cartographic.fromCartesian(position, scene.globe.ellipsoid, new Cartographic()),
-  ]);
-  return sample.height;
 }
